Transforms proxy, tokens and tools with default export
